% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\name{db_snowflake_copy}
\alias{db_snowflake_copy}
\title{Perform a COPY INTO in Snowflake to perform a load or unload operation.}
\usage{
db_snowflake_copy(con, from, to, format_opts = list())
}
\arguments{
\item{con}{A SnowflakeDBConnection object.}

\item{from}{The source of the data, i.e., what comes after the FROM in the COPY
statement. This can be a table, a subquery, a stage, or a local file.}

\item{to}{The target of the COPY statement as a string. This can be a S3/Azure/local
filesystem location, a table, or a Snowflake stage.}

\item{format_opts}{A list of key-value pairs for the Snowflake COPY file_format options.}
}
\description{
Perform a COPY INTO in Snowflake to perform a load or unload operation.
}
\examples{
\dontrun{
# Connection basics ---------------------------------------------------------
# To connect to a database first create a src:
my_db <- src_snowflakedb(user = "snowman",
                         password = "letitsnow",
                         account = "acme",
                         opts = list(warehouse = "mywh",
                                     db = "mydb",
                                     schema = "public")
# Copy from a local CSV file to a target table in my_db.
db_snowflake_copy(my_db$con, from = "file:///my/directory/foo.csv", to = "target_table", 
format_opts = list(format = 'csv', field_delimiter = ','))
}
}
