% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data-lahman.R
\name{lahman}
\alias{lahman}
\title{Cache and retrieve an \code{src_snowflakedb} of the Lahman baseball database.}
\description{
This creates an interesting database using data from the Lahman baseball
data source, provided by Sean Lahman at
\url{http://www.seanlahman.com/baseball-archive/statistics/}, and
made easily available in R through the \pkg{Lahman} package by
Michael Friendly, Dennis Murphy and Martin Monkman. See the documentation
for that package for documentation of the inidividual tables.
}

