% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lahman.R
\name{lahman_snowflakedb}
\alias{lahman_snowflakedb}
\title{Lahman function for SnowflakeDB.}
\usage{
lahman_snowflakedb(...)
}
\arguments{
\item{...}{Parameters to pass through for src_snowflakedb()}
}
\description{
Lahman function for SnowflakeDB.
}
