% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\docType{class}
\name{SnowflakeDBConnection-class}
\alias{SnowflakeDBConnection-class}
\title{dplyr backend support for SnowflakeDB (https://snowflake.net)}
\description{
Use \code{src_snowflakedb} to connect to an existing Snowflake database,
and \code{tbl} to connect to tables within that database.
}
