% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.snowflakedb.R
\docType{package}
\name{dplyr.snowflakedb}
\alias{dplyr.snowflakedb}
\alias{dplyr.snowflakedb-package}
\title{dplyr.snowflakedb: A SnowflakeDB backend for dplyr}
\description{
dplyr.snowflakedb: A SnowflakeDB backend for dplyr
}

